working atax for the manycore baseline version with scalar reduce for now, next is SIMD version

+++ programs-phil/spad/atax/atax.h
@@ -8,7 +8,7 @@
 #define REGION_SIZE (PREFETCH_LEN * 2)
 #define NUM_REGIONS (512 / REGION_SIZE)
 
-typedef float DTYPE;
+typedef int DTYPE;
 
 // pthread argument for the kernel
 typedef struct Kern_Args
